StringField: support undo when inserting text
When text is inserted, user should be able to use ctrl-z to undo the
insertion.

The command needs the input to be focused.
But the developer may want to keep the focus where it is and only insert
text (which is the existing behavior). This is why the focus is
restored after text inserted.

 364018