Menu: move overflown property to action
If an action is overflown, it means that the container is too small to
show the action. This is typically used for action bars (like menu bar)
where the action is put into an ellipsis menu or made hidden in another
way. This is not limited to menu and menu bar, so to make it easier
for custom action bars, the property is moved to Action.