HybridManager: add support for context elements in hybrid events
The "data" attribute on hybrid events can hold static information. But
sometimes, it would be helpful if we could reference a widget (e.g.
Tree) or any other model element (e.g. TreeNode). Since the IDs of
those elements are only known in the JSON layer, we cannot send them
as part of the "data" object

A new property "contextElements" is introduced for that purpose. It can
hold named lists of context elements. A context element consists of a
widget and optionally any other element owned by that widget. In JSON
form, a widget is represented by its JSON adapter ID. The optional
element is represented by a widget-specific ID. These transformations
are automatically applied by the JSON layer using an extensible list
of converters (HybridActionContextElementConverter).

Other notable changes:
[HybridManagerAdapter.ts]
  Rename HybridEvent -> HybridRemoteEvent
[HybridManager.ts]
  Remove deprecated methods
[JsonHybridManagerTest.java]
  Simplify existing test cases. Use DataObjectMapper instance instead
  of custom mock logic. This used to be necessary, because no
  implementation existed at runtime, but that has changed since.
[UiSessionMock.java]
  Correctly register spy instance in the JSON adapter registry.
  Otherwise, resolving an adapter by ID might return an unexpected
  instance (the original adapter instead of the spy).
[scout.ts]
  Add Constructor type.

389630